Refactor: Rename 'vehiclesList' folder to 'vehicles' and update related imports and names.
- Rename folder 'vehiclesList' to 'vehicles'.
- Update vehicles slice name.
- Update all vehicles-related imports.
- Update vehicles reducer name in 'store.js'.
- Update vehicles access in dispatches.
