+++ src/app/store.js
@@ -1,13 +1,13 @@
 import { configureStore } from '@reduxjs/toolkit';
 
 import usersReducer from '../features/users/usersSlice';
-import vehiclesListReducer from '../features/vehiclesList/vehicleListSlice';
+import vehiclesReducer from '../features/vehicles/vehiclesSlice';
 import reservationsReducer from '../features/reservations/reservationsSlice';
 
 export default configureStore({
   reducer: {
     users: usersReducer,
-    vehiclesList: vehiclesListReducer,
+    vehicles: vehiclesReducer,
     reservationsList: reservationsReducer,
   },
 });
