+++ src/features/vehicles/VehiclesList.jsx
@@ -4,7 +4,7 @@ import { useSelector, useDispatch } from 'react-redux';
 import Spinner from 'react-bootstrap/Spinner';
 
 import SplideCarousel from './SplideCarousel';
-import fetchVehicles from './vehicleListThunk';
+import fetchVehicles from './vehiclesThunk';
 
 import '../../styles/features/vehiclesList/VehiclesList.scss';
 
@@ -22,7 +22,7 @@ const VehiclesList = () => {
   }, [reFetchVehicles]);
 
   const { loading, error, vehicles } = useSelector(
-    (state) => state.vehiclesList,
+    (state) => state.vehicles,
   );
 
   if (loading) {
